Add squish method
Since the parser seems to have problems doing is assoc<list> on my
Formula infix operators, this commit adds a squish method which
recursively merges parent and child nodes which have the same operator.
This method is applied now on construction of AST::Operator objects
from the infixes and emulates is assoc<list>.

The listiness tests work now, but right-associativity of ⇒ still not.