Add rewrite engine to AST
AST formulas can now be rewritten according to patterns and substitutions,
which use the smartmatch operator on subtrees. The rewriting rules are
by default applied in order until none could be applied for the first
time. (Note: it may not halt and no heuristics are applied to detect this).
A rule is a pair

  (^:p ⇔ ^:q) => { ($:p ⇒ $:q) ∧ ($:q ⇒ $:p) }

with a formula as key and a Callable as value. ^:p is RewriteCapture object
created via &prefix:<^> from a pair. The rewrite engine tries to find the
left rule key formula in the invocant AST and keeps a hash of subformulas
captured from RewriteCaptures. This hash is passed as named arguments to
the Callable that is the value of the rewrite rule. Its return value
replaces the current matched node. In the example above, a ⇔ node will
be replaced by its definition, which begins with an ∧ node and includes
two copies each of the subtrees at the origin ⇔ node.

In general, a RewriteCapture must have a name which is used in the
substitution Callable and can have a matcher, which is anything that
allows smartmatching against, to restrict which subtrees can go into
the RewriteCapture. To match only ¬ in front of variables, not
arbitrary formulas, use

  (¬^:p(Variable)) => { $:p } # remove ¬ but only in front of variables

Or (:^p(Variable)) => { ¬$:p } will negate all variables, but not the
formula itself. By default :p has value True, which will always succeed
on smartmatch, so ^:p DWIMs.