Make semantics of rewrite on listy operators consistent
rewrite had a fundamental problem with listy operators and operators
that could be listy but aren't. Consider the Boolean formula p ∧ q ∧ r.
Intuitively, it should be matched by the pattern ^:p ∧ ^:r.
Until this commit, whether this intuition holds true was up to chance
(and unlikely). The squished representation of the formula that is
created by default by the infix operators is (∧ p q r) which does
not match (∧ ^p ^q), because part of the matching compares the number
of operands.

If the formula was squished implicitly and if the exact equality
requirement for number of operands were loosened, e.g. to an inequality,
new problems arise. In case only the operands in the smallest of both
lists were compared, we get false positives with short LHS and long
pattern. If the number of operands on the LHS were required to be
at least the number of operands in the pattern, it would pose a technical
difficulty to extract only the matched operands from a matching operator
and leave the other ones in place.

Now, the formula being rewritten as well as all rules are automatically
spread before the rewrite engine starts. This clarifies the semantics
of rewriting listy operators: they are treated as if they were spread,
according to their associativity. Indeed, p ∧ q ∧ r is now guaranteed
to be matched by ^:p ∧ ^:q where ^:p contains (p ∧ q) and ^:q contains
(r). This is because ∧ is left associative. In the future, one might
think about adverbs which allow to locally modify this ordering in
rewrite rules.