Remove Propositional::AST::Variable
As the corresponding FIXME note stated, it is tempting to introduce a
wrapper role around Propositional::Variable which implements AST-specific
functions, like squish, spread and rewrite. However, such an AST-y type
could not be removed from variables when formulas are converted to
non-AST representations such as the Propositional::CNF that will soon
appear. This is unclean, as is adding AST-specific methods to the
top-level Propositional::Variable.

Now, these instances are special-cased in the Operator role: squish and
spread use the safe call syntax .? because they action on variables would
be trivial. rewrite now inspects operands more closely instead of blindly
recursing.

I'm willing to pay that price, because in this module variables are
supposed to be very independent, user-supplied objects that we cannot
manipulate freely. They shouldn't carry around methods from all possible
representations of formulas, and the possibility of introducing specific
subtypes was argued against already above.

The only very unfavorable consequence of this is that you cannot safely
call squish, spread or rewrite on just any random Formula \φ that someone
gives you: it might be just a single variable which doesn't implement
these methods. [ That said, calling squish always works, because it is
a method of Any and coincidentally works as Operator.squish needs it. ]