Commit partial progress on implementing a SAT::Job
I will pursue a different approach in the master branch but want to
keep this idea around.

One main problem that a SAT::Job was supposed to solve comes from the
DIMACS format. It requires variables to be encoded by unique integers.
To make AllSAT and #SAT solvers work, this range must be contiguous
and start at 1 (if you start at 2, there is only unused variable '1'
which will double all your satisfying assignments, for instance).

There must be a DIMACSification step which uses either the variable's
type to establish a correspondence between integers and variable
objects, or, if that's not available for whatever reason [*], an ad-hoc
cache where every newly encountered variable is inserted to number them
uniquely and consistently.

[*] The reason might be that the variables are rather unorganized,
or you don't care about them being organized. It is also possible that
you share variables of the same type in unrelated formulas (like in
tests) and you want a #SAT computation to use only the variables in
the formula at hand, not the variables known to the variable type,
but which are used in other formulas.

Since we also want user-defined filters to act on formulas, we would
want to turn DIMACSification into one of these filters as well, to make
them arbitrarily composable. But now, there is the AllSAT problem.
If you use an ad-hoc cache in the input path to the solver and lose
that cache, you are unable to decode the assignments coming out of it.
A simple $clauses.map.map.map chain is not sufficient if state must
be shared between encoding and decoding. Therefore the idea is to
introduce a SAT::Job class and add en- and decoding layers around its
input and output.

This is a sound idea but the way I picture the code now, the user won't
have to interact much with these jobs and their filters, so their
modularization in classes and roles has lower demand than I initially
thought. I value the flexibility and composability of using maps more,
and the cache sharing problem of DIMACSify is a very special one.
In that case, one can provide a container `is rw` to the encoder to
store its cache in and pass that later to the decoder.

All the abstraction introduced by SAT::Job would also not help solving
a more grave probem with DIMACS: the number of variables and clauses
must be known at the very beginning. The map chain doesn't help with
that either, but it's more lightweight, so I'm pursuing that strategy
on the main branch now.