Spread replacements in rewrite
Correct and consistent execution of rewrite depends on everything,
the patterns and the formula to be spread. This is the case initially
but the rewrite-once method could insert non-spread replacements
or create non-spread subformulas by inserting (even spread) replacements
in the wrong position.

Example: (∧ (∧ p q) r) is the spread version of (∧ p q r). It was
possible for a rewrite rule to insert (∧ p q r) instead because
return values of the rewrite rule were not spread. It was also
possible to rewrite (∧ p q) into (∧ p (∧ a b)) by rewriting q,
which is now not spread either, so pattern matching might fail
in the next iteration.

These two problems broke normal forms (which rely on rewriting until
no rule could be applied anymore) on sufficiently complex formulas.