Fix another cause of too few rewrites
The last commit message contained an example of a situation where a
replacement that is itself spread could result in a non-spread sub-
formula, again making it possible for a rewrite to terminate pre-
maturely because pattern matching failed. This situation was not
in the regression tests and still broken. It is not fixed and has
a test.

The fix is a rather expensive .squish.spread chain after every
rewrite-once call in rewrite. A normalize method should be added
which does the same thing but avoids a full squishing.