Start implementing SAT solver interface
This commit starts work on Propositional::SAT, an interface to SAT
solvers. It converts formulas into DIMACS cnf (a non-trivial task,
see below) and uses the SAT module to solve the SAT, #SAT and AllSAT
problems. Application code will be very small for what it can do.

Some difficulties arise from having to serialize formulas into the
very rudimentary DIMACS cnf (SAT competition) format. We have to map
Variable objects uniquely to integers, for one, which is done by a
Lookup structure, for which a role has to be exposed. A default
implementation using a cache is provided, sufficient for common uses
and reasonably efficient.

Currently, while the Propositional::CNF::Clause input can be a
memory-efficient lazy Seq, their corresponding DIMACS clauses have
to be held in memory until all clauses have been processed.
This is because the DIMACS format expects the number of variables
and clauses in the header. This source of inefficiency should be
fixed at a later time.

Lastly, tautologies apparently cannot really be encoded in DIMACS,
or at least the only candidate for such an encoding, namely

  p cnf $num-vars 1
  0

is not interpreted correctly by all solvers. Tests involving tautologies
are currently skipped and would fail.