Fix SAT routines for tautologies
Since the usual SAT solvers can't handle tautologies (and it is unclear
if DIMACS even supports the encoding of them), this commit brings solvers
made to handle the special case of DIMACS that I declare to mean tautology:

  p cnf $num-vars 0

The number of variables is everything that is required to answer the
varieties of SAT problems about a tautology.

The SAT tests now pass but are capricious: sometimes one of them fails,
and always a different one. Even the type of solvers differs. I don't
know what causes this. The individual SAT solver modules also have
stressful tests and never flapped.